Remove the NotebookControllerFactory service
NotebookControllersLive is the only code that creates Python controllers.
Replace the Effect service with a scoped constructor:

    const controller = yield* createPythonController({
      id,
      label,
      env,
    })

The constructor reads its Effect dependencies directly and returns the same
PythonController. NotebookControllersLive still owns each controller scope,
selection subscription, environment refresh, and disposal. Rename the module
to PythonController to match what it now exports.
