+++ extension/src/kernel/__tests__/NotebookRuntimeOperations.test.ts
@@ -18,8 +18,8 @@ import { TestTelemetryLive } from "../../__mocks__/TestTelemetry.ts";
 import { NotebookRange, TestVsCode } from "../../__mocks__/TestVsCode.ts";
 import { makeTestMarimoClient } from "../../__tests__/__utils__/TestMarimoClient.ts";
 import { NOTEBOOK_TYPE, SCRATCH_CELL_ID } from "../../constants.ts";
-import { PythonController } from "../../kernel/NotebookControllerFactory.ts";
 import { NotebookRuntime } from "../../kernel/NotebookRuntime.ts";
+import { PythonController } from "../../kernel/PythonController.ts";
 import { cellId } from "../../lib/__tests__/branded.ts";
 import { VsCode } from "../../platform/VsCode.ts";
 import { MarimoNotebookDocument } from "../../schemas/MarimoNotebookDocument.ts";
